(PC-24596)[PRO] refactor: rename Template to IndivualOfferLayout
