+++ pro/src/pages/IndividualOfferWizard/Confirmation/Confirmation.tsx
@@ -2,16 +2,16 @@ import React from 'react'
 
 import { useIndividualOfferContext } from 'context/IndividualOfferContext'
 import { IndividualOffer } from 'core/Offers/types'
-import { Template as WizardTemplate } from 'screens/IndividualOffer'
+import IndivualOfferLayout from 'screens/IndividualOffer/IndivualOfferLayout/IndivualOfferLayout'
 import { IndividualOfferConfirmation as IndividualOfferConfirmationScreen } from 'screens/IndividualOfferConfirmation'
 
 const Confirmation = (): JSX.Element => {
   const { offer } = useIndividualOfferContext()
   // FIXME : we should not need  as IndividualOffer cause parent route would redirect on offer loading error
   return (
-    <WizardTemplate withStepper={false}>
+    <IndivualOfferLayout withStepper={false}>
       <IndividualOfferConfirmationScreen offer={offer as IndividualOffer} />
-    </WizardTemplate>
+    </IndivualOfferLayout>
   )
 }
