+++ pro/src/pages/IndividualOfferWizard/PriceCategoriesSummary/PriceCategoriesSummary.tsx
@@ -1,13 +1,13 @@
 /* istanbul ignore file: DEBT, TO FIX */
 import React from 'react'
 
-import { Template as WizardTemplate } from 'screens/IndividualOffer'
+import IndivualOfferLayout from 'screens/IndividualOffer/IndivualOfferLayout/IndivualOfferLayout'
 import { PriceCategoriesSummaryScreen } from 'screens/IndividualOffer/PriceCategoriesSummary/PriceCategoriesSummary'
 
 export const PriceCategoriesSummary = (): JSX.Element | null => {
   return (
-    <WizardTemplate title="Récapitulatif">
+    <IndivualOfferLayout title="Récapitulatif">
       <PriceCategoriesSummaryScreen />
-    </WizardTemplate>
+    </IndivualOfferLayout>
   )
 }
