+++ pro/src/screens/IndividualOffer/IndivualOfferLayout/IndivualOfferLayout.module.scss
