+++ pro/src/screens/IndividualOffer/IndivualOfferLayout/OfferStatusBanner/OfferStatusBanner.tsx
