+++ pro/src/screens/IndividualOffer/IndivualOfferLayout/OfferStatusBanner/index.ts
