--- pro/src/screens/IndividualOffer/Template/index.ts
@@ -1,1 +1,0 @@
-export { default as Template } from './Template'
