+++ pro/src/screens/IndividualOffer/index.ts
@@ -2,5 +2,3 @@ export { Informations } from './Informations'
 export { StocksThing } from './StocksThing'
 export { StocksEventEdition } from './StocksEventEdition'
 export { Summary } from './Summary'
-export { Template } from './Template'
-export { PriceCategories } from './PriceCategories/PriceCategories'
